Генерация тегов на уровне проекта
