+++ lib/dapp.rb
@@ -53,6 +53,7 @@ require 'dapp/dapp/git_artifact'
 require 'dapp/dapp/dappfile'
 require 'dapp/dapp/chef'
 require 'dapp/dapp/dapp_config'
+require 'dapp/dapp/option_tags'
 require 'dapp/dapp/logging/base'
 require 'dapp/dapp/logging/process'
 require 'dapp/dapp/logging/paint'
@@ -286,7 +287,6 @@ require 'dapp/dimg/image/stage'
 require 'dapp/dimg/image/scratch'
 require 'dapp/dimg/dimg/git_artifact'
 require 'dapp/dimg/dimg/path'
-require 'dapp/dimg/dimg/tags'
 require 'dapp/dimg/dimg/stages'
 require 'dapp/dimg/dimg'
 require 'dapp/dimg/artifact'
